Moves public audit types to core
