+++ packages/core/security/core-security-server/index.ts
@@ -8,11 +8,15 @@
 
 export type { SecurityServiceSetup, SecurityServiceStart } from './src/contracts';
 export type { CoreAuthenticationService } from './src/authc';
+export type { CoreAuditService } from './src/audit';
 export type {
   CoreSecurityDelegateContract,
   AuthenticationServiceContract,
+  AuditServiceContract,
 } from './src/api_provider';
 export type {
   SecurityRequestHandlerContext,
   AuthcRequestHandlerContext,
 } from './src/request_handler_context';
+export type { AuditEvent, AuditHttp, AuditKibana, AuditRequest } from './src/audit/audit_events';
+export type { AuditLogger } from './src/audit/audit_logger';
