+++ packages/core/security/core-security-server/src/api_provider.ts
@@ -6,6 +6,7 @@
  * Side Public License, v 1.
  */
 
+import { CoreAuditService } from './audit';
 import type { CoreAuthenticationService } from './authc';
 
 /**
@@ -16,9 +17,12 @@ import type { CoreAuthenticationService } from './authc';
  */
 export interface CoreSecurityDelegateContract {
   authc: AuthenticationServiceContract;
+  audit: AuditServiceContract;
 }
 
 /**
  * @public
  */
 export type AuthenticationServiceContract = CoreAuthenticationService;
+
+export type AuditServiceContract = CoreAuditService;
