+++ packages/core/security/core-security-server/src/audit.ts
@@ -1,15 +1,16 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
  */
 
 import type { KibanaRequest } from '@kbn/core/server';
 
-import type { AuditLogger } from './audit_logger';
+import type { AuditLogger } from './audit/audit_logger';
 
-export interface AuditServiceSetup {
+export interface CoreAuditService {
   /**
    * Creates an {@link AuditLogger} scoped to the current request.
    *
