+++ x-pack/packages/security/plugin_types_server/src/audit/index.ts
@@ -6,5 +6,10 @@
  */
 
 export type { AuditServiceSetup } from './audit_service';
-export type { AuditEvent, AuditHttp, AuditKibana, AuditRequest } from './audit_events';
-export type { AuditLogger } from './audit_logger';
+export type {
+  AuditEvent,
+  AuditHttp,
+  AuditKibana,
+  AuditLogger,
+  AuditRequest,
+} from '@kbn/core-security-server';
