+++ x-pack/packages/security/plugin_types_server/tsconfig.json
@@ -14,5 +14,6 @@
     "@kbn/core",
     "@kbn/security-plugin-types-common",
     "@kbn/core-user-profile-server",
+    "@kbn/core-security-server",
   ]
 }
